pipeline/requirements.txt
pandas
pandas-gbq
db-dtypes
xgboost
shap
scikit-learn
matplotlib
joblib
pytest
pytest-cov
pre-commit
flask
isort
black
flake8
python-dotenv
beautifulsoup4
html5lib
lxml
requests
psutil
tqdm
pyarrow
selenium
webdriver-manager
pyyaml
shap>=0.40.0
numba>=0.54.0
fastapi
uvicorn
httpx
python-multipart
feedparser
tabulate
praw
google-cloud-bigquery
pydantic
